IPRT: RT_MORE_STRICT for r0rdv and r0drv/darwin.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@22052 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/r0drv/darwin/RTLogWriteDebugger-r0drv-darwin.cpp
@@ -33,13 +33,13 @@
 *   Header Files                                                               *
 *******************************************************************************/
 #include "the-darwin-kernel.h"
+#include "internal/iprt.h"
 #include <iprt/log.h>
-#include <iprt/assert.h>
 
 
 RTDECL(void) RTLogWriteDebugger(const char *pch, size_t cb)
 {
-    kprintf("%.*s", cb, pch);
+    kprintf("%.*s", (int)cb, pch);
     return;
 }
